Add separate component for selected reef card
